+++ src/sys/socket/sockopt.rs
@@ -492,6 +492,15 @@ sockopt_impl!(
     libc::LOCAL_PEERCRED,
     super::XuCred
 );
+#[cfg(any(target_os = "macos", target_os = "ios"))]
+sockopt_impl!(
+    /// Get the PID of the peer process of a connected unix domain socket.
+    LocalPeerPid,
+    GetOnly,
+    0,
+    libc::LOCAL_PEERPID,
+    libc::c_int
+);
 #[cfg(any(target_os = "android", target_os = "linux"))]
 sockopt_impl!(
     /// Return the credentials of the foreign process connected to this socket.